+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Correction in the use of the api with mongodb
 * Adjustments to search endpoint for contacts, chats, messages and Status messages
 * Now when deleting the instance, the data referring to it in mongodb is also deleted
+* It is now validated if the instance name contains uppercase and special characters
 
 # 1.3.1 (2023-07-20 07:48)